lit+macaroons: move BakeSuperMacaroons helper
Move this helper to the new macaroons package.
